Fixed tests and linter

+++ power/src/app/fragebogen/dashboard/dashboard.component.ts
@@ -55,6 +55,11 @@ export class DashboardComponent implements OnInit {
     }
 
     public async deleteForm(id: string) {
+        // Ask user to confirm deletion
+        if (!confirm($localize`Möchten Sie dieses Formular wirklich löschen?`)) {
+            return;
+        }
+
         try {
             this.loadingscreen.setVisible(true);
             const response = await this.formAPI.deleteInternForm(id);
